+++ src/main/scala/dev/template/spark/Main.scala
@@ -3,8 +3,6 @@ package dev.template.spark
 import dev.template.spark.source.Reader
 import org.apache.spark.sql.functions.avg
 
-import java.io.File
-
 case class Person(firstName: String, lastName: String, country: String, age: Int)
 
 class CalculateAverageAge extends SparkSessionWrapper with Reader with Logger {
@@ -30,18 +28,21 @@ class CalculateAverageAge extends SparkSessionWrapper with Reader with Logger {
 
 object Main extends App {
   if (args.length == 0) {
-    println(""" USAGE :
-              | spark-submit \
-              | --class dev.template.spark.Main \
-              | --packages io.delta:delta-core_2.12:2.4.0 \
-              | --master spark://localhost:7077 \
-              | --deploy-mode client \
-              | --driver-memory 1g \
-              | --executor-memory 1g \
-              | --executor-cores 2 \
-              | build/libs/spark-scala-gradle-bootstrap-2.12.0-all.jar \
-              | src/main/resources/people-example.csv
-              |""".stripMargin)
+    println(
+      """ USAGE :
+        | Running in local mode local[2]
+        |
+        | ${SPARK_HOME}/bin/spark-submit \                                                                                                                                          ✔  6s   
+        |                --verbose  \
+        |                --class dev.template.spark.Main \
+        |                --packages io.delta:delta-core_2.12:2.4.0 \
+        |                --master "local[2]" \
+        |                --driver-memory 1g \
+        |                --executor-memory 1g \
+        |                --executor-cores 2 \
+        |                build/libs/spark-scala-gradle-bootstrap-2.12.0-all.jar \
+        |                src/main/resources/people-example.csv
+        |""".stripMargin)
     throw new RuntimeException("Requires input file people-example.csv")
   }
   private val inputFilePath = args(0)