+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 First release version of the project
 
 ### Added
+
 - support for gradle 7.6, java 11, spark 3.4.1, scala 2.12 and delta 2.4.0
 - support to read file:// and s3://
 - Added CHANGELOG.md
@@ -18,3 +19,10 @@ First release version of the project
 - Added tests for the above classes
 - Added scalaCodeCoverage
 - Added spotless, checkstyle, and spotbugs
+
+## [1.0.1] - 2024-02-17
+
+### Added
+
+- Added Wordcount example with GCS dataproc configuration and steps to run to README.md
+- Deleted unwanted files from resources